% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelefreqanalysis.R
\name{cervus_alf}
\alias{cervus_alf}
\title{Allele Frequency Analysis}
\usage{
cervus_alf(
  cervusCL_location,
  analysis_folder,
  analysis_name = "CERVUS_Analysis",
  genotype_file,
  NLoci,
  DoHardyWeinberg = TRUE,
  HWMinExpectedFrequency = 5,
  UseYatesCorrection = TRUE,
  UseBonferroniCorrection = TRUE,
  DoNullAllele = TRUE
)
}
\arguments{
\item{cervusCL_location}{Path to cervusCL.exe (CERVUS is found at http://www.fieldgenetics.com)}

\item{analysis_folder}{Path to the folder which contains the files required for the analysis. All output will also be saved in this folder.}

\item{analysis_name}{Can specify a custom name for the analysis, which will be saved as the .crv filename}

\item{genotype_file}{The filename of the genotype file. Currently, the function expects a .csv file with the first column as the id, then the loci starting in column two. Headers are assumed and loci names are retained.}

\item{NLoci}{Number of loci in the genotype file}

\item{DoHardyWeinberg}{TRUE/FALSE Do Hardy-Weinberg test box to test each locus for whether it conforms to HW equilibrium}

\item{HWMinExpectedFrequency}{Default is 5}

\item{UseYatesCorrection}{TRUE/FALSE Use Yates correction when calculating chi-square values with one degree of freedom}

\item{UseBonferroniCorrection}{TRUE/FALSE Use Bonferroni correction when carrying out the HW test for multiple loci simultaneously}

\item{DoNullAllele}{TRUE/FALSE Estimate null allele frequency}
}
\value{
Currently the function creates the .crv file that contains the settings for the analysis, then performs the analysis. The results are saved in the usual CERVUS format, and printed to the console.
}
\description{
Perform an Allele Frequency Analysis uisng CervusCL
}
\examples{
cervus_alf(cervusCL_location = cervusCL_location,
           analysis_folder = analysis_folder,
           genotype_file = "genotype_file.csv",
           analysis_name = "test_analysis",
           NLoci = 6,
           DoHardyWeinberg = TRUE,
           HWMinExpectedFrequency = 5,
           UseYatesCorrection = TRUE,
           UseBonferroniCorrection = TRUE,
           DoNullAllele = TRUE
           )

}
