% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CervusPAR.R
\name{CervusPAR}
\alias{CervusPAR}
\title{Run parentage analysis using CervusCL}
\usage{
CervusPAR(
  CervusCLPath,
  AnalysisFolderPath,
  AnalysisName = "Cervus_Analysis",
  AnalysisType,
  OffspringFile_FileName,
  OffspringFile_HasHeader = TRUE,
  OffspringFile_OffspringIDColumnNumber = 1,
  OffspringFile_IncludesKnownParents = FALSE,
  OffspringFile_KnownParentIDColumnNumber = 0,
  OffspringFile_IncludesCandidateParents = FALSE,
  OffspringFile_CandidateParentIDColumnNumber = 0,
  CandidateFemaleFile_FileName = NULL,
  CandidateFemaleFile_HasHeader = NULL,
  CandidateFemaleFile_CandidateParentFormat = "One column for all offspring",
  CandidateFemaleFile_OffspringIDColumnNumber = NULL,
  CandidateFemaleFile_CandidateParentIDColumnNumber = NULL,
  CandidateMaleFile_FileName = NULL,
  CandidateMaleFile_HasHeader = NULL,
  CandidateMaleFile_CandidateParentFormat = "One column for all offspring",
  CandidateMaleFile_OffspringIDColumnNumber = NULL,
  CandidateMaleFile_CandidateParentIDColumnNumber = NULL,
  UseSimulationParameters = TRUE,
  CalculateConfidenceLevels = TRUE,
  AlwaysTestSelfing = FALSE,
  MinTypedLoci,
  UseCorrectedLikelihoods = TRUE,
  UseMistypingRateAsLikelihoodErrorRate = TRUE,
  LikelihoodErrorRate = 0.01,
  CriticalStatisticName = "Delta",
  TruncateAtZero = TRUE,
  OutputType = "The most-likely parent",
  OutputSortedBy = "Joint LOD score",
  OutputIncludeNonExclusionProbabilities = FALSE
)
}
\arguments{
\item{CervusCLPath}{Path to cervusCL.exe (Cervus is found at http://www.fieldgenetics.com).}

\item{AnalysisFolderPath}{Path to the folder which contains the files required for the analysis. All output will also be saved in this folder.}

\item{AnalysisName}{Can specify a custom name for the analysis, which will be saved in the output filenames.}

\item{AnalysisType}{Set as "Maternity"  or "Paternity" or "Parent pair (sexes known)" or "Parent pair (sexes unknown)".}

\item{OffspringFile_FileName}{Name of the file with offspring data}

\item{OffspringFile_HasHeader}{TRUE/FALSE: Does the file have a header?}

\item{OffspringFile_OffspringIDColumnNumber}{Column number with offspring ID}

\item{OffspringFile_IncludesKnownParents}{TRUE/FALSE}

\item{OffspringFile_KnownParentIDColumnNumber}{Column number with offspring ID (set to 0 if no known parents)}

\item{OffspringFile_IncludesCandidateParents}{TRUE/FALSE: does the offspring file also contain candidate parents?}

\item{OffspringFile_CandidateParentIDColumnNumber}{Column number where candidate parent IDs start (set to 0 if no candidate parents in the file)}

\item{CandidateFemaleFile_FileName}{Name of the file with candidate females}

\item{CandidateFemaleFile_HasHeader}{TRUE/FALSE: Does the file have a header?}

\item{CandidateFemaleFile_CandidateParentFormat}{"One column for all offspring"}

\item{CandidateFemaleFile_OffspringIDColumnNumber}{If using a format other than "One column for all offspring", give the column number that contains the offspring ID}

\item{CandidateFemaleFile_CandidateParentIDColumnNumber}{If using a format other than "One column for all offspring", give the column number that contains the first candidate parent ID}

\item{CandidateMaleFile_FileName}{Name of the file with candidate males}

\item{CandidateMaleFile_HasHeader}{TRUE/FALSE: Does the file have a header?}

\item{CandidateMaleFile_CandidateParentFormat}{"One column for all offspring"}

\item{CandidateMaleFile_OffspringIDColumnNumber}{If using a format other than "One column for all offspring", give the column number that contains the offspring ID}

\item{CandidateMaleFile_CandidateParentIDColumnNumber}{If using a format other than "One column for all offspring", give the column number that contains the first candidate parent ID}

\item{UseSimulationParameters}{TRUE/FALSE: TRUE (default) to use the SIM parameters}

\item{CalculateConfidenceLevels}{TRUE/FALSE: TRUE (default)}

\item{AlwaysTestSelfing}{TRUE/FALSE: Should selfing be tested?}

\item{MinTypedLoci}{Minimum number of typed loci to be included. Should match that of the simulation.}

\item{UseCorrectedLikelihoods}{TRUE/FALSE: Use Kalinowski et al. (2007) likelihood equations (recommended)}

\item{UseMistypingRateAsLikelihoodErrorRate}{TRUE/FALSE}

\item{LikelihoodErrorRate}{Default is 0.01}

\item{CriticalStatisticName}{Delta" or "LOD". "Delta" recommended. See the Cervus help files.}

\item{TruncateAtZero}{TRUE/FALSE}

\item{OutputType}{Default is "The most-likely parent", see the Cervus help files for more.}

\item{OutputSortedBy}{Default is "Joint LOD score", see the Cervus help files for more.}

\item{OutputIncludeNonExclusionProbabilities}{TRUE/FALSE}
}
\value{
Currently prints the results to the console, and saves the parentage analysis data in AnalysisFolderPath
}
\description{
You should be familiar with Cervus before using this function. 
The arguments relate directly to settings within Cervus, usually selected by the user using the GUI.
Argument names are such that it can easily be inferred what setting they change within a Cervus project file (.crv).
}
