% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CervusSIM.R
\name{CervusSIM}
\alias{CervusSIM}
\title{Run parentage simulations using CervusCL}
\usage{
CervusSIM(
  CervusCLPath,
  AnalysisFolderPath,
  AnalysisName = "CervusAnalysis",
  AnalysisType,
  NOffspring = 10000,
  NCandidateMales = 0,
  PropCandidateMalesSampled = 0,
  NCandidateFemales = 0,
  PropCandidateFemalesSampled = 0,
  PropLociTyped,
  PropLociMistyped = 0.01,
  MinTypedLoci,
  CriticalStatisticName = "Delta",
  TruncateAtZero = TRUE,
  RelaxedConfidence = 80,
  StrictConfidence = 95,
  SimulateInbreeding = FALSE,
  ParentRelatedness = 0,
  InbreedingRate = 0,
  AlwaysTestSelfing = FALSE,
  SimulateFemaleRelatives = FALSE,
  FemalePropRelatives = 0,
  FemaleRelatedTo = "Offspring",
  FemaleRelatedness = 0,
  SimulateMaleRelatives = FALSE,
  MalePropRelatives = 0,
  MaleRelatedTo = "Offspring",
  MaleRelatedness = 0,
  UseCorrectedLikelihoods = TRUE,
  UseMistypingRateAsLikelihoodErrorRate = TRUE,
  LikelihoodErrorRate = 0.01,
  RepeatSimulation = FALSE,
  NRepeats = 1,
  ApplyPreviousSimulationData = FALSE,
  GenerateTables = FALSE,
  SaveRawStatisticScores = FALSE,
  GenerateHistograms = FALSE,
  NCategories = 5,
  MinStatistic = 0,
  MaxStatistic = 0,
  PreviousSimulationDataFile = "",
  ResultsToConsole = TRUE,
  wineCommand = NA,
  wineHomeDirectory = "Z:",
  wineTempDirectory = NA
)
}
\arguments{
\item{CervusCLPath}{Path to cervusCL.exe (Cervus is found at http://www.fieldgenetics.com).}

\item{AnalysisFolderPath}{Path to the folder which contains the files required for the analysis. All output will also be saved in this folder.}

\item{AnalysisName}{Can specify a custom name for the analysis, which will be saved in the output filenames.}

\item{AnalysisType}{Set as "Maternity"  or "Paternity" or "Parent pair (sexes known)" or "Parent pair (sexes unknown)".}

\item{NOffspring}{Number of offspring to simulate, default to 10000, but use more for final analysis.}

\item{NCandidateMales}{Average number of candidate male parents per offspring, regardless of being sampled (set to 0 for maternity analysis).}

\item{PropCandidateMalesSampled}{Proportion of candidate male parents that are sampled (1 is not recommended, set to 0 for maternity analysis).}

\item{NCandidateFemales}{Average number of candidate female parents per offspring, regardless of being sampled (set to 0 for paternity analysis).}

\item{PropCandidateFemalesSampled}{Proportion of candidate female parents that are sampled (1 is not recommended, set to 0 for paternity analysis).}

\item{PropLociTyped}{Proportion of loci typed allows for missing data and should be an average value across loci and individuals. It is estimated in the ALF.}

\item{PropLociMistyped}{Proportion of loci mistyped allows for errors in genotyping and should be an average value across loci and individuals. See the Cervus help files for how it handles errors.}

\item{MinTypedLoci}{Threshold for number of typed loci for an individual to be included in the analysis.}

\item{CriticalStatisticName}{"Delta" or "LOD". "Delta" recommended. See the Cervus help files.}

\item{TruncateAtZero}{TRUE/FALSE}

\item{RelaxedConfidence}{Relaxed confidence level (default is 80\%)}

\item{StrictConfidence}{Strict confidence level (default is 95\%)}

\item{SimulateInbreeding}{TRUE/FALSE: Should inbreeding be simulated? You should also set ParentRelatedness and InbreedingRate if TRUE.}

\item{ParentRelatedness}{The probability that an allele in one parent is identical by descent to the corresponding allele in the other parent.}

\item{InbreedingRate}{The proportion of simulated offspring whose true parents are related to each other.}

\item{AlwaysTestSelfing}{TRUE/FALSE: Test for self-fertilisation}

\item{SimulateFemaleRelatives}{TRUE/FALSE: simulate relatives of the mother-father-offspring trio among the pool of candidate parents.}

\item{FemalePropRelatives}{The average proportion of female candidate parents that are relatives.}

\item{FemaleRelatedTo}{Relatives can be related to the offspring, its true mother or its true father. See Cervus helpfiles.}

\item{FemaleRelatedness}{The probability that each allele belonging to a relative is identical by descent to the corresponding allele carried by the individual to which they are related.}

\item{MalePropRelatives}{TRUE/FALSE: simulate relatives of the mother-father-offspring trio among the pool of candidate parents.}

\item{MaleRelatedTo}{Relatives can be related to the offspring, its true mother or its true father. See Cervus helpfiles.}

\item{MaleRelatedness}{The probability that each allele belonging to a relative is identical by descent to the corresponding allele carried by the individual to which they are related.}

\item{UseCorrectedLikelihoods}{TRUE/FALSE: Use Kalinowski et al. (2007) likelihood equations (recommended)}

\item{UseMistypingRateAsLikelihoodErrorRate}{TRUE/FALSE}

\item{LikelihoodErrorRate}{Default is 0.01}

\item{RepeatSimulation}{TRUE/FALSE: Repeat simulation using the same parameters}

\item{NRepeats}{How many repeats?}

\item{ApplyPreviousSimulationData}{TRUE/FALSE: See Cervus helpfiles.}

\item{GenerateTables}{TRUE/FALSE: See Cervus helpfiles.}

\item{SaveRawStatisticScores}{TRUE/FALSE: See Cervus helpfiles.}

\item{GenerateHistograms}{TRUE/FALSE: See Cervus helpfiles.}

\item{NCategories}{Histogram setup}

\item{MinStatistic}{Histogram setup}

\item{MaxStatistic}{Histogram setup}

\item{PreviousSimulationDataFile}{Path to previous simulation data file (not required unless ApplyPreviousSimulationData = TRUE)}

\item{ResultsToConsole}{TRUE/FALSE: Should the results of the allele frequency analysis be output into the console?}

\item{wineCommand}{ONLY FOR WINE USERS: give the command that should be used to call your installed version of wine (e.g. wine64)}

\item{wineHomeDirectory}{ONLY FOR WINE USERS: Where should Wine find your home directory? This is usually "Z:".}

\item{wineTempDirectory}{ONLY FOR WINE USERS: Point wine to your Windows temp folder (e.g. "/Users/myname/.wine/drive_c/users/myname/Temp"). This is to fix a wine bug.}
}
\value{
Currently prints the results to the console, and saves the simulation data in AnalysisFolderPath
}
\description{
You should be familiar with Cervus before using this function. 
The arguments relate directly to settings within Cervus, usually selected by the user using the GUI.
Argument names are such that it can easily be inferred what setting they change within a Cervus project file (.crv), often at the expense of brevity. This may change in future versions.
This function is wine compatible (see docs).
}
