% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCervusALF.R
\name{ImportCervusALF}
\alias{ImportCervusALF}
\title{Import summary of allele frequency analysis}
\usage{
ImportCervusALF(ALFSummaryFile)
}
\arguments{
\item{CervusALFSummaryFile}{The path to the allele frequency analysis summary file (.txt file)}
}
\value{
Returns the tables in the Cervus allele frequency analysis summary file, formatted for use in R
}
\description{
Imports the summary file produced from a Cervus allele frequency analysis (the .txt file). Can be called directly in the CervusALF function with ImportALF = TRUE.
}
\examples{
ImportCervusALF(ALFSummaryFile = "AlleleFrequencyAnalysis.txt")
}
