% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCervusPAR.R
\name{ImportCervusPAR}
\alias{ImportCervusPAR}
\title{Import summary and data from a Cervus parentage analysis}
\usage{
ImportCervusPAR(PARSummaryFile, PARResultsFile, known_parents = TRUE)
}
\arguments{
\item{PARSummaryFile}{The summary file (.txt)}

\item{PARResultsFile}{The results file (.sim)}

\item{known_parents}{Was the mother known? Default TRUE}
}
\value{
A list containing the summarised information, and the results of the parentage analysis.
}
\description{
Imports the summary and data files produced from a Cervus parentage analysis.
}
