% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelefreqanalysis.R
\name{CervusALF}
\alias{CervusALF}
\title{Allele Frequency Analysis}
\usage{
CervusALF(
  CervusCLPath,
  AnalysisFolderPath,
  AnalysisName = "CERVUS_Analysis",
  GenotypeFile_FileName,
  GenotypeFile_HasHeader = TRUE,
  GenotypeFile_ReadLocusNames = TRUE,
  GenotypeFile_IDColumnNumber = 1,
  GenotypeFile_FirstAlleleColumnNumber = 2,
  GenotypeFile_ColumnsPerLocus = 2,
  GenotypeFile_NLoci,
  DoHardyWeinberg = TRUE,
  HWMinExpectedFrequency = 5,
  UseYatesCorrection = TRUE,
  UseBonferroniCorrection = TRUE,
  DoNullAllele = TRUE
)
}
\arguments{
\item{CervusCLPath}{Path to cervusCL.exe (Cervus is found at http://www.fieldgenetics.com)}

\item{AnalysisFolderPath}{Path to the folder which contains the files required for the analysis. All output will also be saved in this folder.}

\item{AnalysisName}{Can specify a custom name for the analysis, which will be saved as the .crv filename}

\item{GenotypeFile_FileName}{The filename of the genotype file. (e.g. genotype.csv)}

\item{GenotypeFile_HasHeader}{TRUE/FALSE: Does the genotype file contain a header row}

\item{GenotypeFile_ReadLocusNames}{TRUE/FALSE: Should Cervus read the locus names in the file (TRUE), or give them generic names (FALSE)}

\item{GenotypeFile_IDColumnNumber}{Column number that identifies each individual}

\item{GenotypeFile_FirstAlleleColumnNumber}{Column number which contains the first allele}

\item{GenotypeFile_ColumnsPerLocus}{What format are the loci in? One allele per column = 2 columns per loci. See Cervus helpfiles for more info.}

\item{GenotypeFile_NLoci}{Number of loci in the genotype file}

\item{DoHardyWeinberg}{TRUE/FALSE: Do Hardy-Weinberg test box to test each locus for whether it conforms to HW equilibrium}

\item{HWMinExpectedFrequency}{Default is 5}

\item{UseYatesCorrection}{TRUE/FALSE: Use Yates correction when calculating chi-square values with one degree of freedom}

\item{UseBonferroniCorrection}{TRUE/FALSE: Use Bonferroni correction when carrying out the HW test for multiple loci simultaneously}

\item{DoNullAllele}{TRUE/FALSE Estimate null allele frequency}
}
\value{
Currently the function creates the .crv file that contains the settings for the analysis, then performs the analysis. The results are saved in the usual CERVUS format, and printed to the console.
}
\description{
Perform an Allele Frequency Analysis uisng CervusCL
}
\examples{
CervusALF(
  CervusCLPath = "C:\\\\Program Files (x86)\\\\Field Genetics\\\\Cervus\\\\Cervus CL\\\\CervusCL.exe",
  AnalysisFolderPath = "C:\\\\Analysis",
  AnalysisName = "CERVUS_Analysis",
  GenotypeFile_FileName = "genotype_file.csv",
  GenotypeFile_HasHeader = TRUE,
  GenotypeFile_ReadLocusNames = TRUE,
  GenotypeFile_IDColumnNumber = 1,
  GenotypeFile_FirstAlleleColumnNumber = 2,
  GenotypeFile_ColumnsPerLocus = 2,
  GenotypeFile_NLoci = 6,
  DoHardyWeinberg = TRUE,
  HWMinExpectedFrequency = 5,
  UseYatesCorrection = TRUE,
  UseBonferroniCorrection = TRUE,
  DoNullAllele = TRUE
}
